% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{calculate_model_fit}
\alias{calculate_model_fit}
\title{Fit ARIMA model}
\usage{
calculate_model_fit(df, model_order, drift = FALSE)
}
\arguments{
\item{df}{data.frame with nested list column of time-series}

\item{model_order}{ARIMA model parameters}

\item{drift}{ARIMA model drift parameter}
}
\value{
original list of successfully fitted models
}
\description{
Fit ARIMA model
}
