% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sex_meta_data.R
\name{calculate_sex_meta_data}
\alias{calculate_sex_meta_data}
\title{Calculate sex meta-analysis data}
\usage{
calculate_sex_meta_data(data = clean$hbc)
}
\arguments{
\item{data}{high-burden country data}
}
\description{
Calculate sex meta-analysis data
}
