% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{compare_observed_predicted}
\alias{compare_observed_predicted}
\title{Evaluate model prediction}
\usage{
compare_observed_predicted(observed, predicted, fn)
}
\arguments{
\item{observed}{data frame of observed notifications}

\item{predicted}{data frame of predicted notifications}

\item{fn}{function to compare values}
}
\description{
Evaluate model prediction
}
