% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{evaluate_predictions}
\alias{evaluate_predictions}
\title{Predict and compare model predictions}
\usage{
evaluate_predictions(
  observed_data = observed,
  order = c(0, 1, 0),
  drift = FALSE,
  evaluation_function
)
}
\arguments{
\item{observed_data}{data frame of observed notifications}

\item{order}{model specification}

\item{drift}{model drift parameter}

\item{evaluation_function}{function to evaluate observed and predicted
data - requires tow arguments (observed and predicted)}
}
\description{
Predict and compare model predictions
}
