% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{extract_model_fit}
\alias{extract_model_fit}
\title{Extract model fit}
\usage{
extract_model_fit(models, param)
}
\arguments{
\item{models}{list of models}

\item{param}{model parameter to be extracted and added}
}
\value{
vector of doubles
}
\description{
Extract model fit
}
