% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{forecast_next_year}
\alias{forecast_next_year}
\title{Forecast following year}
\usage{
forecast_next_year(tbl, order, drift, lambda = NULL, log_transform = FALSE)
}
\arguments{
\item{tbl}{data frame including specific nested list columns}

\item{order}{parameters to define ARIMA model}

\item{drift}{drift parameter for ARIMA model}

\item{lambda}{Box-Cox transformation parameter. If lambda="auto",
then a transformation is automatically selected using BoxCox.lambda.
The transformation is ignored if NULL.
Otherwise, data transformed before model is estimated.}

\item{log_transform}{logical to log transform count time series}
}
\description{
Forecast following year
}
