% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_predictions.R
\name{grouped_predictions}
\alias{grouped_predictions}
\title{Gruoped predictions}
\usage{
grouped_predictions(prediction_df, ...)
}
\arguments{
\item{prediction_df}{data frame containing prediction point
estimates and standard deviations}

\item{...}{grouping variables}
}
\description{
Aggregate predictions across grouping variables
}
\examples{
\dontrun{
grouped_predictions(df, type, age_group)
}
}
