% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_case_counts.R
\name{impute_missing_case_counts}
\alias{impute_missing_case_counts}
\title{Impute missing counts}
\usage{
impute_missing_case_counts(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Impute missing counts
}
