% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{prepare_actual_notifications}
\alias{prepare_actual_notifications}
\title{Prepare observed notifications}
\usage{
prepare_actual_notifications(original_data, year = 2019, training_data)
}
\arguments{
\item{original_data}{list of data frames}

\item{year}{numeric to define which year to restrict to}

\item{training_data}{training data set}
}
\description{
Combine cleaned data frames, restrict to specific year and merge with
training data sets
}
