% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{prepare_data}
\alias{prepare_data}
\title{Create long and wide data sets}
\usage{
prepare_data(raw, const = constants())
}
\arguments{
\item{raw}{data frame of raw notification data}

\item{const}{analysis constants}
}
\value{
list
}
\description{
Create long and wide data sets
}
