% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{prepare_long_data}
\alias{prepare_long_data}
\title{Prepare long analysis data}
\usage{
prepare_long_data(lst, const)
}
\arguments{
\item{lst}{list of data frames representing clean notification data}

\item{const}{analysis constants}
}
\description{
Prepare long analysis data
}
