% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{render_rmd}
\alias{render_rmd}
\title{Create reports}
\usage{
render_rmd(
  rmd_filename,
  output_dir = path.expand("~"),
  save_artefacts = FALSE,
  plot_type = c("png", "svg")
)
}
\arguments{
\item{rmd_filename}{string: name of Rmd file}

\item{output_dir}{string: optional path to output directory. Defaults
to "~" directory}

\item{save_artefacts}{boolean: define whether plots and tables should be
saved as separate files in the output_dir}

\item{plot_type}{string: png or svg}
}
\description{
Create reports
}
