% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stringency_index}
\alias{stringency_index}
\title{Stringency index}
\format{
A data frame with 186 rows and 820 variables:
}
\usage{
stringency_index
}
\description{
A global dataset documenting daily stringency index values during the 
COVID-19 panedemic
}
\keyword{datasets}
