% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb_notifications}
\alias{tb_notifications}
\title{WHO country TB data}
\format{
A data frame with 8707 rows and 198 variables:
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
tb_notifications
}
\description{
A global dataset containing TB notification data
}
\keyword{datasets}
